# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Revendis API",
    "version": "0.2.0",
    "description": "Initial OpenAPI contract for Sprint 1 endpoints.",
    "license": {
      "name": "Proprietary",
      "identifier": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://api.revendis.com",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/inventory/products": {
      "get": {
        "summary": "List products",
        "operationId": "listProducts",
        "tags": [
          "inventory"
        ],
        "responses": {
          "200": {
            "description": "Products list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      },
      "post": {
        "summary": "Create product",
        "operationId": "createProduct",
        "tags": [
          "inventory"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/inventory/adjustments": {
      "post": {
        "summary": "Adjust inventory",
        "operationId": "adjustInventory",
        "tags": [
          "inventory"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryAdjustmentInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Adjustment applied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/inventory/transfers": {
      "post": {
        "summary": "Transfer inventory between stores",
        "operationId": "transferInventory",
        "tags": [
          "inventory"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryTransferInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transfer completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/inventory/returns": {
      "post": {
        "summary": "Register return",
        "operationId": "registerReturn",
        "tags": [
          "inventory"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryReturnInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Return registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object"
          }
        }
      },
      "ListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
      "ItemResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object"
          }
        }
      },
      "ProductInput": {
        "type": "object",
        "required": [
          "name",
          "sku",
          "price"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "barcode": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "cost": {
            "type": "number"
          },
          "stock": {
            "type": "number"
          }
        }
      },
      "InventoryAdjustmentInput": {
        "type": "object",
        "required": [
          "sku",
          "quantity",
          "reason"
        ],
        "properties": {
          "sku": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "reason": {
            "type": "string"
          },
          "storeId": {
            "type": "string"
          }
        }
      },
      "InventoryTransferInput": {
        "type": "object",
        "required": [
          "sku",
          "quantity",
          "fromStoreId",
          "toStoreId"
        ],
        "properties": {
          "sku": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "fromStoreId": {
            "type": "string"
          },
          "toStoreId": {
            "type": "string"
          }
        }
      },
      "InventoryReturnInput": {
        "type": "object",
        "required": [
          "saleId",
          "items"
        ],
        "properties": {
          "saleId": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "sku": {
                  "type": "string"
                },
                "quantity": {
                  "type": "number"
                }
              }
            }
          },
          "condition": {
            "type": "string",
            "enum": [
              "good",
              "damaged"
            ]
          }
        }
      },
...
    }
  }
}